Allow top-level 'action' and 'last-action', matching all emails
I also considered making this a match expression as

match all {
    <actions>
}

but having it as a top-level seems to match closer with how Outlook represents
the trivial filter: namely by having an empty list of properties. Since this
list is usually of length <the number of nested match blocks>, having an empty
list corresponds naturally to having no nesting.

More evidence for this being the best approach is that this is a 4-line patch.